+++ src/common/scalar.utils.ts
@@ -1,0 +1,3 @@
+import { Int } from '@nestjs/graphql';
+
+export const returnsInt = () => Int;