+++ src/connection/index.ts
@@ -1,0 +1,5 @@
+export * from './args';
+export * from './connection.type';
+export * from './args/forward-connection-args.type';
+export * from './page-info.type';
+export * from './resolve-connection-field.decorator';