+++ src/global-object-identification/index.ts
@@ -1,0 +1,9 @@
+export {
+  GlobalId,
+  GlobalIdFieldOptions,
+  GlobalIdFieldResolver,
+  GlobalIdScalar,
+  ResolvedGlobalId,
+} from './global-id';
+
+export { NodeFieldResolver, NodeInterface, NodeType } from './node';