+++ src/global-object-identification/node/node-field.resolver.ts
@@ -1,10 +1,6 @@
 import { Query, Args, Resolver } from '@nestjs/graphql';
 import { returnsNodeInterface, returnsNodeInterfaces, NodeInterface } from './node.interface';
-import {
-  ResolvedGlobalId,
-  typeResolvedGlobalId,
-  typeResolvedGlobalIds,
-} from './resolved-global-id.class';
+import { ResolvedGlobalId, typeResolvedGlobalId, typeResolvedGlobalIds } from '../global-id';
 
 export type ResolvedNode =
   | Promise<NodeInterface>