+++ src/mutation/utils/capitalise.util.ts
@@ -1,0 +1,1 @@
+export const capitalise = (text: string): string => text.charAt(0).toUpperCase() + text.slice(1);