+++ test/starwars-app/ship.service.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@nestjs/common';
-import { FactionService } from './faction.service';
 
 export interface ShipDTO {
   id: number;
@@ -8,8 +7,6 @@ export interface ShipDTO {
 
 @Injectable()
 export class ShipService {
-  constructor(private factionService: FactionService) {}
-
   private ships: ShipDTO[] = [
     {
       id: 1,
@@ -51,12 +48,9 @@ export class ShipService {
     return this.ships.find((ship) => ship.id === id);
   }
 
-  public createShip(name: string, factionId: string) {
+  public createShip(name: string) {
     const id = this.nextShipId++;
     this.ships.push({ id, name });
-
-    this.factionService.assignShip(factionId, id);
-
     return this.getShip(id);
   }
 }